Tutorial 8
First look at while loops and I wrote a for loop for the last task
(supplementary).

Firstly there is a simple while loop that does some basic arithmetic.
Next there is tool for tutors to input marks and saved to .txt file and
the corresponding grades are printed to the screen.

I used a do while loop that loops round as many times as there are
students and has some user validation to prevent against invalid
percentages. I then wrote a while loop that reads from the file that was
just created and converts the percentages into grades

Finally, I used a for loop in the last task to find the largest factor
of any number smaller than 32 bits of memory. I could change the program
to use the long data type for greater versatility.