Tutorial 11
Look at variable length arrays. There are various classes of these but I
focused on ArrayList. The only shortcomings of this type of array are
that they are not as efficient as fixed sized arrays and they can't be
used with primitive data types.

In this tutorial I work with the basic methods of ArrayList to add and
remove elements etc.